+++ BestRoom/House.h
@@ -6,12 +6,12 @@ using namespace std;
 
 struct House {
     string roadNameAddress; //도로명 주소
-    double monthly;     // 월세 + 관리비
-    double deposit;      // 보증금
-    double distance;    // 신공학관으로부터 거리
-    double size;         // 평수
+    float monthly;     // 월세 + 관리비
+    float deposit;      // 보증금
+    float distance;    // 신공학관으로부터 거리
+    float size;         // 평수
     string url;
-    double totalScore;
+    float totalScore;
 };
 #endif
